requirements.txt
fastapi[standard]==0.114.0
pydantic==2.9.0
redis==5.0.8
pytest==8.3.2
python-dotenv==1.0.1